Move dakota.dakota to dakota.core
Because it's weird using "from dakota.dakota import Dakota". Although
easy to remember.
