+++ dakota/bmi_dakota.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 """Basic Model Interface for the Dakota iterative systems analysis toolkit."""
 
-from .dakota import Dakota
+from .core import Dakota
 
 
 class BmiDakota(object):
